+++ bin/static/app.js
@@ -3065,4 +3065,4 @@
   };
 })();
 /* ui-bundle: 22b2db40599f2f84a3a63059485494dbffecd6f6c86fc435edf9d5a8c736404e */
-/* ui-sources: b7a8fd3a260b7239f6e3980e8467fa5beed7d61e93c6fb1f5fccfe7d40aedcd1 */
+/* ui-sources: 535214e67ba28ebfc64e890d524a0e9633bffa377d5b7028f34903323f2e703c */